requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
algoliasearch==3.0.0
algoliasearch-django==3.0.0
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2024.2
cryptography==43.0.1
cssbeautifier==1.15.1
Django==5.1.1
django-allauth==64.2.1
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-filebrowser==4.0.3
django-grappelli==3.0.9
django-tinymce==4.1.0
djangorestframework==3.15.2
djlint==1.35.2
easy-thumbnails==2.10
EditorConfig==0.12.4
Faker==28.4.1
frozenlist==1.4.1
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.9
jsbeautifier==1.15.1
json5==0.9.25
jwt==1.3.1
kombu==5.4.1
multidict==6.1.0
pathspec==0.12.1
pillow==10.4.0
prompt_toolkit==3.0.47
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.11.1
